fix(recommendation): correct repository generic type from Integer to Long

+++ src/main/java/com/hyewon/grocey_api/domain/recommendation/repository/RecipeRecommendationRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RecipeRecommendationRepository extends JpaRepository<RecipeRecommendation, Integer> {
+public interface RecipeRecommendationRepository extends JpaRepository<RecipeRecommendation, Long> {
     List<RecipeRecommendation> findByUser(User user);
     List<RecipeRecommendation> findByFridge(Fridge fridge);
     void deleteByUser(User user);
